Altered Touch Paint - Interrupt Pin Capable
Modified from POLLING to interrupt processing of Touch.  When PIN
interrupt is available it sleeps until the pin is active then returns
data during touch when user app requests it.  If the PIN interrupt is
not specified or working - it runs a Timer1 interrupt to POLL in the
background for the user and will present data when available.

Added debug subwindow

This is a prototype - more examples to be updated with Interrupt Pin
support and better use of STRUCT'ured Button handling.