INTERRUPT - No SPI POLLING version of OnOffButton2
I redid the interrupt code outside the BUTTON handler just aliased the
Touch and Point calls and made my adjustments there.

I see the same occasional missed/lost interrupts as in my prior code
that was tightly bound in the Button Code.  I'm not sure how to detect
this - checking the PIN2 value for LOW==TOUCH gives ODD results.

This code will use a TIMER or PIN interrupt to prevent polling based on
the rate of the user call.  I only tested specifying BOTH this time -
when the pin interrupt hits it will disable the timer that effectively
polls at a steady rate.  Once an interrupt is reported the user will be
free to POLL as the desired rate until two requests for TOUCH return
NO_TOUCH, then polling is disabled until the next interrupt detection. I
had it try once again after first NO_TOUCH to assure that one failure to
read wasn't detected while a touch was incoming.

Most often the interrupt failure is not seeing the touch and allowing
polling to begin?  A test for that would be a second button interrupt to
trigger the code to read when a continuous touch is being ignored.

There are 6 buttons of 4 types on the screen.  Below that is a user
feedback area showing the button reported and busy prints for ongoing
activity in some fashion, there is also USB recording of the buttons,
but easier to watch the screen as you touch.