Problem found and fixable now in the library code
A pin interrupt will trigger given the slightest touch.  There is a
threshold value above that for a valid point.  The problem I was seeing
previously where a touch starts light and is held, was that the
interrupt associated with the light touch was wasted and then the
reading was disabled so the user finger was touching and the one time
interrupt was ignored, where subsequent readings find and return  valid
touch readings.. This change disabled the interrupt poll allowed flag to
a lower level.

This goes back to what I saw where the timer based interrupt was always
fine - because it would reset some milliseconds later and find the touch
in progress.

I need to test on the LC again - as its slower response time may be what
was behind the higher failure rate there.  My more complex earlier code
was trying to fix this symptom by manually doing another read after the
light touch was seen.  Not only did this complicate the process - but
the problem still persisted with the high threshold causing the
secondary read to fail as well.